third_party/blink/web_tests/virtual/user-activation-v2/fullscreen/full-screen-iframe-zIndex-expected.txt

CONSOLE WARNING: line 20: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
